# blueprint/openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele Wedding Registry API
  description: >
    Pool-funded wedding gift registry platform with kinship-tier gift
    suggestions, multi-currency support, and ACID-safe contributions.
  version: "1.0.0"
  contact:
    name: Saukele Support
    email: support@saukele.kz
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token. Obtain via POST /auth/login.
        Include as: Authorization: Bearer <token>
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        code:
          type: string
          nullable: true
          example: VALIDATION_ERROR
        error:
          type: string
          example: Resource not found
        message:
          type: string
          nullable: true
          example: Validation failed
        details:
          type: object
          nullable: true
        requestId:
          type: string
          nullable: true

    # ── Monetary precision-safe scalars ───────────────────────────────────
    MoneyAmount:
      type: string
      pattern: '^\\d+(\\.\\d{1,2})?$'
      example: "15000.00"
      description: Decimal monetary amount represented as string to avoid floating-point rounding.
    ExchangeRate:
      type: string
      pattern: '^\\d+(\\.\\d{1,6})?$'
      example: "497.368421"
      description: Decimal exchange rate represented as string with up to 6 fractional digits.

    # ── Pagination cursor ──────────────────────────────────────────────────
    CursorMeta:
      type: object
      properties:
        nextCursor:
          type: integer
          nullable: true
          example: 42
          description: Pass as ?cursor= in next request. Null means no more pages.
        take:
          type: integer
          example: 20
        hasMore:
          type: boolean
          example: true

    # ── User ──────────────────────────────────────────────────────────────
...
    GiftItemResponse:
      type: object
      properties:
        id:
          type: integer
          example: 7
        registryId:
          type: integer
          example: 3
        title:
          type: string
          example: KitchenAid Stand Mixer
        description:
          type: string
          example: 5-quart artisan stand mixer in Empire Red
        imageUrl:
          type: string
          nullable: true
          example: https://cdn.saukele.kz/items/kitchenaid.jpg
        targetAmountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          example: "189000.00"
        currentAmountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          example: "75600.00"
        targetAmountEur:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          nullable: true
          example: "380.00"
        exchangeRateAtTime:
          allOf:
            - $ref: '#/components/schemas/ExchangeRate'
          nullable: true
          example: "497.368421"
          description: KZT/EUR rate locked at item creation
        lockedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-01-20T08:00:00.000Z"
        status:
          type: string
          enum: [PENDING, FUNDED, PURCHASED, DELIVERED]
          example: PENDING
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
          example: HIGH
        createdAt:
          type: string
          format: date-time
...
    ContributionCreate:
      type: object
      required: [amountKzt, contributorName]
      properties:
        amountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          description: Must be >= 1000.00 KZT
          example: "15000.00"
        contributorName:
          type: string
          example: Darkhan Seitkali
        message:
          type: string
          example: Бақыт тілеймін!
    ContributionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 22
        giftItemId:
          type: integer
          example: 7
        userId:
          type: integer
          nullable: true
          example: 5
        contributorName:
          type: string
          example: Darkhan Seitkali
        amountKzt:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          example: "15000.00"
        amountEur:
          allOf:
            - $ref: '#/components/schemas/MoneyAmount'
          nullable: true
          example: "30.17"
        exchangeRateAtTime:
          allOf:
            - $ref: '#/components/schemas/ExchangeRate'
          example: "497.368421"
          description: KZT/EUR rate locked at contribution time (immutable)
        lockedAtTimestamp:
          type: string
          format: date-time
          example: "2026-02-01T14:30:00.000Z"
        message:
          type: string
          nullable: true
          example: Бақыт тілеймін!
        status:
          type: string
          enum: [PENDING, CONFIRMED, REFUNDED]
          example: CONFIRMED
        createdAt:
          type: string
          format: date-time

    # ── FamilyMember ──────────────────────────────────────────────────────
...
    KinshipTierResponse:
      type: object
      properties:
        kinshipTier:
          type: string
          enum: [CLOSE, EXTENDED, FRIEND]
          nullable: true
          example: EXTENDED
        suggestedMinKzt:
          type: integer
          example: 20000
...
paths:
...
  /items/{itemId}/contributions:
    post:
      tags: [Contributions]
      summary: Contribute to a pool-funded gift item
      description: >
        Auth optional. If authenticated, userId is recorded.
        Exchange rate is fetched from ExchangeRateSnapshot and locked immutably.
        Runs inside a Prisma $transaction for ACID safety.
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
          example: 7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionCreate'
      responses:
        '201':
          description: Contribution recorded; returns updated gift item state
          content:
            application/json:
              schema:
                type: object
                properties:
                  contribution:
                    $ref: '#/components/schemas/ContributionResponse'
                  giftItem:
                    $ref: '#/components/schemas/GiftItemResponse'
                  kinshipSuggestion:
                    $ref: '#/components/schemas/KinshipTierResponse'
        '400':
          description: Amount below 1000 KZT or exceeds remaining target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Gift item not accepting contributions (PURCHASED/DELIVERED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Contributions]
      summary: List contributions for a gift item (owner or ADMIN; cursor-based pagination)
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: integer
        - name: take
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REFUNDED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt]
            default: "-createdAt"
      responses:
        '200':
          description: Paginated contributions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContributionResponse'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the registry owner and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Gift item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contributions/my:
    get:
      tags: [Contributions]
      summary: List contributions created by the authenticated user (ADMIN can filter by userId)
      security:
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: integer
        - name: take
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REFUNDED]
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt]
            default: "-createdAt"
      responses:
        '200':
          description: Paginated contribution list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContributionResponse'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contributions/{id}:
    get:
      tags: [Contributions]
      summary: Get a single contribution by ID (owner, contributor, or ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Contribution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Contribution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── FAMILY MEMBERS ────────────────────────────────────────────────────────
...